+++ src/Main.kt
@@ -1,3 +1,7 @@
+import problem.*
+
 fun main() {
     println("Welcome to kopt")
+    val problem = Creator[BinCover]?.invoke("basic", 1)
+    println(problem)
 }