+++ src/discrete/Constraint.kt
@@ -10,8 +10,8 @@ typealias ConstraintFn = (Solution) -> Boolean
 class Constraint(
 	val test: ConstraintFn,
 	val penalty: Penalty,
-	val variables: List<Variable>,
-	){
+	val variables: List<Variable>
+){
 	fun isSatisfied(solution: Solution) = test(solution)
 	fun computePenalty(solution: Solution): Penalty = if (isSatisfied(solution)) 0.0 else penalty
 }